% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.loess.R
\name{plot.loess}
\alias{plot.loess}
\title{Scatter plot with LOESS line}
\usage{
\method{plot}{loess}(x, y = NULL, loess.col = "red", ...)
}
\arguments{
\item{x}{Either a data frame with two columns, which will be taken as x and y, or x coordinates}

\item{y}{If x is not a data frame, y coordinates}
}
\description{
This function makes an x-y scatter plot and adds a LOESS line
}
\examples{
plot.loess(cars)
}

