% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotloess.R
\name{plotloess}
\alias{plotloess}
\title{Scatter plot with LOESS line}
\usage{
plotloess(x, y = NULL, loess.col = "red", ...)
}
\arguments{
\item{x}{Either a data frame with two columns, which will be taken as x and y, or x coordinates}

\item{y}{If x is not a data frame, y coordinates}

\item{loess.col}{The colour of the LOESS line}

\item{...}{Other arguments passed to plot}
}
\description{
This function makes an x-y scatter plot and adds a LOESS line. Used by sim.residplot.
}
\examples{
plotloess(cars)
}

