% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residfitted.olre.R
\name{residfitted.olre}
\alias{residfitted.olre}
\title{Output "corrected" Pearson residuals and fitted values}
\usage{
residfitted.olre(object, warn = TRUE)
}
\arguments{
\item{object}{Fitted model object}

\item{warn}{Warn when not Poisson glmer object with OLRE}
}
\description{
Output "corrected" Pearson residuals and fitted values from a Poisson
GLMM with an observation-level random effect (OLRE) fitted using lme4::glmer.
Defaults to residuals(., type = "pearson") and fitted(.) for other classes of fit.
}
\examples{
library(lme4)
form1 <- TICKS ~ YEAR + scale(HEIGHT) + (1 | BROOD) + (1 | LOCATION) + (1 | INDEX)
fit.poisln  <- glmer(form1, family = "poisson", data = grouseticks)
residfitted.olre(fit.poisln)
}

